requirements.txt
avro-python3==1.10.2
certifi==2021.10.8
charset-normalizer==2.0.12
confluent-kafka==1.8.2
fastavro==1.4.10
idna==3.3
psycopg2-binary==2.9.3
requests==2.27.1
urllib3==1.26.9
